requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
requests
python-dotenv
openai
msal